requirements1.txt
cython
numpy>=1.11.3
scipy>=0.19.1
imbalanced-learn>=0.3.0
ipdb
IPython>=6.2.1
Jinja2>=2.9.6
joblib>=0.12.5
lmfit>=0.9.11
marshmallow>=3.0.0b4
matplotlib>=2.1.0
networkx>=2.0
nilearn>=0.3.0
pandas>=0.20.3
python-igraph>=0.7.1.post6
scikit-learn>=0.19.0
seaborn>=0.7.1
statsmodels>=0.8.0
svgutils>=0.3.0
jupyter
